♻️ refactor: consolidate user router and add type safety for navigation
